Separated the form into a new component and kept concerns dealing with the login page separate from login form client component
