+++ src/app/auth/login/page.module.scss
@@ -27,40 +27,4 @@
     opacity: 0.8;
     height: 100vh;
   }
-
-  .formContainer {
-    display: flex;
-    flex-direction: column;
-    padding: 15px;
-    position: relative;
-    margin: 0 auto;
-    top: 250px;
-    left: 450px;
-    width: 30%;
-    border-radius: 20px;
-    @include flexibleBackgroundColor(147, 177, 193, 0.8);
-  }
-  
-  .inputContainer {
-    display: flex;
-    justify-content: center;
-    flex-direction: column;
-  }
-  
-  .input {
-    background-color: transparent;
-    border: 2px solid gray;
-  }
-  
-  h1 {
-    position: relative;
-    text-align: center;
-  }
-  label {
-    position: relative;
-  } 
-  .authBtn {
-    z-index: 1000;
-    margin-bottom: 20px;
-  }
 }
