+++ src/components/loginForm/login.module.scss
@@ -1,0 +1,36 @@
+@import "../../styles/mixins.module.scss";
+.formContainer {
+  display: flex;
+  flex-direction: column;
+  padding: 15px;
+  position: relative;
+  margin: 0 auto;
+  top: 250px;
+  left: 450px;
+  width: 30%;
+  border-radius: 20px;
+  @include flexibleBackgroundColor(147, 177, 193, 0.8);
+
+  .inputContainer {
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
+  }
+
+  .input {
+    background-color: transparent;
+    border: 2px solid gray;
+  }
+
+  h1 {
+    position: relative;
+    text-align: center;
+  }
+  label {
+    position: relative;
+  }
+  .authBtn {
+    z-index: 1000;
+    margin-bottom: 20px;
+  }
+}
